requirements.txt
asgiref==3.2.10
astroid==2.9.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.13
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pillow==9.0.0
pylint==2.5.3
python-decouple==3.4
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
stripe~=2.64.0
toml==0.10.1
urllib3==1.26.5
wrapt==1.12.1

celery~=5.2.3
djangorestframework~=3.13.1